Track the ISS using Node-RED
In this lab, you will track the live location of the International Space Station (ISS) and connect it to an external service (locationIQ) plus the Watson Assistant, so you can ask the app where over the world is the ISS in real-time. This zip file contains the document, the JSON flow for both the Node-RED flow and the Watson Assistant.